[Feat] Lv 12 랭킹 구현 완료

+++ src/main/java/com/gamebasic/ranking/dto/RankingSource.java
@@ -59,7 +59,7 @@ public class RankingSource {
             private String seed;
             private GameStatus status;
             private int clearedFloor;
-            private long durationSeconds;
+            private int durationSeconds;
             private int finalHp;
             private List<Floor> floors;
